Use float32 for rebin speedup 1300 ms to 1100 ms

+++ python/src/scipp/plot/plot_2d.py
@@ -460,7 +460,7 @@ class Slicer2d(Slicer):
             ],
                              values=self.vslice[slicex][slicey].values,
                              unit=sc.units.counts,
-                             dtype=sc.dtype.float64))
+                             dtype=sc.dtype.float32))
 
         # Also include the masks
         if self.params["masks"][self.name]["show"]:
